Do not pass through isSupportedEngine
It is unnecessary to pass through isSupportedEngine, instead compute it
in checkAppCompatibility when it is actually needed.
