+++ src/launcher/models/index.js
@@ -48,7 +48,6 @@ const ImmutableApp = Record({
     iconPath: null,
     shortcutIconPath: null,
     isOfficial: null,
-    isSupportedEngine: null,
     sharedVersion: null,
     source: null,
     url: null,
@@ -70,7 +69,6 @@ function getImmutableApp(app) {
         iconPath: app.iconPath,
         shortcutIconPath: app.shortcutIconPath,
         isOfficial: app.isOfficial,
-        isSupportedEngine: app.isSupportedEngine,
         source: app.source,
         sharedVersion: app.sharedVersion,
         url: app.url,
