+++ src/launcher/reducers/__tests__/appsReducer-test.js
@@ -55,7 +55,6 @@ const app1 = {
     engineVersion: null,
     homepage: null,
     isOfficial: null,
-    isSupportedEngine: null,
     latestVersion: null,
     sharedVersion: null,
     shortcutIconPath: null,
@@ -75,7 +74,6 @@ const app2 = {
     engineVersion: null,
     homepage: null,
     isOfficial: null,
-    isSupportedEngine: null,
     latestVersion: null,
     sharedVersion: null,
     shortcutIconPath: null,
